+++ apps/remix/.env.example
@@ -8,11 +8,13 @@ TURNSTILE_KEY="1x00000000000000000000AA"
 TURNSTILE_SECRET="1x0000000000000000000000000000000AA"
 
 # turso 中获取, 连接数据库凭证
-TURSO_DB_RO_AUTH_TOKEN=""
+TURSO_DB_RO_AUTH_TOKEN=
 
 # turso 中获取, 连接数据库地址
-TURSO_DB_URL=""
+TURSO_DB_URL=
 # 域名邮箱的域名
-DOMAIN=""
+DOMAIN=
 
 CLOUDFLARE_TURNSTILE_VERIFY_Endpoint="https://challenges.cloudflare.com/turnstile/v0/siteverify"
+
+TURNSTILE_ENABLED=true
