+++ apps/remix/app/components/mail/components/NotSelected.tsx
@@ -1,0 +1,7 @@
+export function NotSelected() {
+  return (
+    <div className="p-8 text-center text-muted-foreground">
+      No message selected
+    </div>
+  );
+}
