+++ apps/remix/app/components/mail/components/account-switcher.tsx
@@ -22,71 +22,65 @@ import {
   DialogTitle,
   DialogTrigger,
 } from "~/components/ui/dialog";
+import { Separator } from "~/components/ui/separator";
+
 interface AccountSwitcherProps {
   isCollapsed: boolean;
   accounts: Account[];
-  siteKey: string;
 }
 
 export function AccountSwitcher({
   isCollapsed,
   accounts,
-  siteKey,
 }: AccountSwitcherProps) {
   const [selectedAccount, setSelectedAccount] = React.useState<string>(
     accounts[0]?.email || ""
   );
 
   return (
-    <>
-      {accounts.length <= 0 ? (
-        <CreateAccount siteKey={siteKey} />
-      ) : (
-        <Select value={selectedAccount} onValueChange={setSelectedAccount}>
-          <SelectTrigger
-            className={cn(
-              "flex items-center gap-2 [&>span]:line-clamp-1 [&>span]:flex [&>span]:w-full [&>span]:items-center [&>span]:gap-1 [&>span]:truncate [&_svg]:h-4 [&_svg]:w-4 [&_svg]:shrink-0",
-              isCollapsed &&
-                "flex h-9 w-9 shrink-0 items-center justify-center p-0 [&>span]:w-auto [&>svg]:hidden"
-            )}
-            aria-label="Select account"
-          >
-            <SelectValue placeholder="Select an email account">
-              <span className={cn("ml-2", isCollapsed && "hidden")}>
-                {
-                  accounts.find((account) => account.email === selectedAccount)
-                    ?.email
-                }
-              </span>
-            </SelectValue>
-          </SelectTrigger>
-          <SelectContent>
-            {accounts.map((account) => (
-              <SelectItem key={account.id} value={account.email}>
-                <div
-                  className="flex items-center gap-3 [&_svg]:h-4 [&_svg]:w-4 [&_svg]:shrink-0 [&_svg]:text-foreground"
-                  key={account.id}
-                >
-                  {account.email}
-                </div>
-              </SelectItem>
-            ))}
-            <CreateAccount siteKey={siteKey} />
-          </SelectContent>
-        </Select>
-      )}
-    </>
+    <Select value={selectedAccount} onValueChange={setSelectedAccount}>
+      <SelectTrigger
+        className={cn(
+          "flex items-center gap-2 [&>span]:line-clamp-1 [&>span]:flex [&>span]:w-full [&>span]:items-center [&>span]:gap-1 [&>span]:truncate [&_svg]:h-4 [&_svg]:w-4 [&_svg]:shrink-0",
+          isCollapsed &&
+            "flex h-9 w-9 shrink-0 items-center justify-center p-0 [&>span]:w-auto [&>svg]:hidden"
+        )}
+        aria-label="Select account"
+      >
+        <SelectValue placeholder="Select an email account">
+          <span className={cn("ml-2", isCollapsed && "hidden")}>
+            {
+              accounts.find((account) => account.email === selectedAccount)
+                ?.email
+            }
+          </span>
+        </SelectValue>
+      </SelectTrigger>
+      <SelectContent>
+        {accounts.map((account) => (
+          <SelectItem key={account.id} value={account.email}>
+            <div
+              className="flex items-center gap-3 [&_svg]:h-4 [&_svg]:w-4 [&_svg]:shrink-0 [&_svg]:text-foreground"
+              key={account.id}
+            >
+              {account.email}
+            </div>
+          </SelectItem>
+        ))}
+        <Separator className="my-2" />
+        <CreateAccount />
+      </SelectContent>
+    </Select>
   );
 }
 
 interface CreateAccountProps {
-  siteKey?: string;
   className?: string;
 }
 
 export function CreateAccount(props: CreateAccountProps) {
   const navigation = useNavigation();
-  const { siteKey = "" } = props;
+
   const [open, setOpen] = useState(false);
   return (
     <Dialog open={open} onOpenChange={setOpen}>
@@ -110,7 +104,7 @@ export function CreateAccount(props: CreateAccountProps) {
         <Form method="POST" className="flex flex-col gap-2 text-center">
           <Turnstile
             className="w-full flex justify-center"
-            siteKey={siteKey}
+            siteKey={"1x00000000000000000000AA"}
             options={{
               theme: "auto",
             }}
