+++ apps/remix/app/components/mail/components/mail-list.tsx
@@ -4,6 +4,7 @@ import { cn } from "~/lib/utils";
 import { ScrollArea } from "~/components/ui/scroll-area";
 import { Mail } from "../data";
 import { useMail } from "../use-mail";
+import { Link } from "@remix-run/react";
 
 interface MailListProps {
   items: Mail[];
@@ -16,26 +17,18 @@ export function MailList({ items }: MailListProps) {
     <ScrollArea className="h-screen">
       <div className="flex flex-col gap-2 p-4 pt-0">
         {items.map((item) => (
-          <button
+          <Link
             key={item.id}
+            to={`/${item.id}`}
             className={cn(
               "flex flex-col items-start gap-2 rounded-lg border p-3 text-left text-sm transition-all hover:bg-accent",
               mail.selected === item.id && "bg-muted"
             )}
-            onClick={() =>
-              setMail({
-                ...mail,
-                selected: item.id,
-              })
-            }
           >
             <div className="flex w-full flex-col gap-1">
               <div className="flex items-center">
                 <div className="flex items-center gap-2">
                   <div className="font-semibold">{item.from.name}</div>
-                  {/* {!item.read && (
-                    <span className="flex h-2 w-2 rounded-full bg-blue-600" />
-                  )} */}
                 </div>
                 <div
                   className={cn(
@@ -55,7 +48,7 @@ export function MailList({ items }: MailListProps) {
             <div className="line-clamp-2 text-xs text-muted-foreground">
               {item.text.substring(0, 300)}
             </div>
-          </button>
+          </Link>
         ))}
       </div>
     </ScrollArea>
