+++ apps/remix/app/components/mail/components/mail.tsx
@@ -1,20 +1,4 @@
-import * as React from "react";
-import { Search } from "lucide-react";
-
-import { Input } from "~/components/ui/input";
-import {
-  ResizableHandle,
-  ResizablePanel,
-  ResizablePanelGroup,
-} from "~/components/ui/resizable";
-import { Separator } from "~/components/ui/separator";
-import { Tabs, TabsContent } from "~/components/ui/tabs";
-import { TooltipProvider } from "~/components/ui/tooltip";
-import { useMail } from "~/components/mail/use-mail";
 import type { Account, Mail } from "../data";
-import { MailList } from "./mail-list";
-import { MailDisplay } from "./mail-display";
-import { AccountSwitcher } from "./account-switcher";
 
 interface MailProps {
   accounts: Account[];
@@ -22,7 +6,6 @@ interface MailProps {
   defaultLayout: number[] | undefined;
   defaultCollapsed?: boolean;
   navCollapsedSize: number;
-  siteKey: string;
 }
 
 export function Mail({
@@ -30,57 +13,6 @@ export function Mail({
   mails,
   defaultLayout = [265, 440, 655],
   defaultCollapsed = false,
-  siteKey,
 }: MailProps) {
-  const [isCollapsed] = React.useState(defaultCollapsed);
-  const [mail] = useMail();
-
-  return (
-    <>
-      <TooltipProvider delayDuration={0}>
-        <ResizablePanelGroup
-          direction="horizontal"
-          onLayout={(sizes: number[]) => {
-            document.cookie = `react-resizable-panels:layout=${JSON.stringify(
-              sizes
-            )}`;
-          }}
-          className="h-full max-h-[860px] items-stretch"
-        >
-          <ResizablePanel defaultSize={defaultLayout[1]} minSize={30}>
-            <Tabs defaultValue="all">
-              <div className="flex items-center px-4 py-2">
-                <AccountSwitcher
-                  isCollapsed={isCollapsed}
-                  accounts={accounts}
-                  siteKey={siteKey}
-                />
-              </div>
-              <Separator />
-              <div className="bg-background/95 p-4 backdrop-blur supports-[backdrop-filter]:bg-background/60">
-                <form>
-                  <div className="relative">
-                    <Search className="absolute left-2 top-2.5 h-4 w-4 text-muted-foreground" />
-                    <Input placeholder="Search" className="pl-8" />
-                  </div>
-                </form>
-              </div>
-              <TabsContent value="all" className="m-0">
-                <MailList items={mails} />
-              </TabsContent>
-              <TabsContent value="unread" className="m-0">
-                <MailList items={mails.filter((item) => !item.read)} />
-              </TabsContent>
-            </Tabs>
-          </ResizablePanel>
-          <ResizableHandle withHandle />
-          <ResizablePanel defaultSize={defaultLayout[2]}>
-            <MailDisplay
-              mail={mails.find((item) => item.id === mail.selected) || null}
-            />
-          </ResizablePanel>
-        </ResizablePanelGroup>
-      </TooltipProvider>
-    </>
-  );
+  return <></>;
 }
