+++ apps/remix/app/config/env.ts
@@ -1,0 +1,1 @@
+export const TURNSTILE_ENABLED = process.env.TURNSTILE_ENABLED;
