+++ apps/remix/app/service/mail.ts
@@ -1,0 +1,8 @@
+export async function fetchMails() {
+  try {
+    const resp = await fetch("/api/mails");
+    return await resp.json();
+  } catch (e) {
+    return [];
+  }
+}
