+++ apps/remix/package.json
@@ -71,7 +71,8 @@
     "tailwindcss-animate": "^1.0.7",
     "tiny-invariant": "^1.3.3",
     "vaul": "^0.9.0",
-    "zod": "^3.22.4"
+    "zod": "^3.22.4",
+    "@tanstack/react-query": "^5.28.8"
   },
   "devDependencies": {
     "@remix-run/dev": "^2.8.1",
