+++ package.json
@@ -23,13 +23,5 @@
     "prettier": "^3.2.5",
     "prettier-plugin-tailwindcss": "^0.5.12",
     "vitepress": "^1.0.1"
-  },
-  "dependencies": {
-    "@libsql/client": "0.4.0-pre.7",
-    "@tanstack/react-query": "^5.28.8",
-    "date-fns": "^3.6.0",
-    "drizzle-orm": "^0.29.5",
-    "drizzle-zod": "^0.5.1",
-    "zod": "^3.22.4"
   }
 }
