--- packages/database/.env.example
@@ -1,10 +1,0 @@
-# cookie的加密 secret, 随机字符串即可
-COOKIES_SECRET=""
-# cloudflare 中获取, 用于网站验证码
-TURNSTILE_KEY=""
-# turso 中获取, 连接数据库使用
-TURNSTILE_SECRET=""
-# turso 中获取, 连接数据库凭证
-TURSO_DB_RO_AUTH_TOKEN=""
-# turso 中获取, 连接数据库地址
-TURSO_DB_URL=""
