+++ packages/database/db.ts
@@ -3,8 +3,8 @@ import { drizzle, LibSQLDatabase } from "drizzle-orm/libsql";
 
 export function getWebTursoDBFromEnv(): LibSQLDatabase {
   const client = createWebClient({
-    url: process.env.TURSO_DB_URL || "",
-    authToken: process.env.TURSO_DB_AUTH_TOKEN || "",
+    url: process.env.TURSO_DB_URL as string,
+    authToken: process.env.TURSO_DB_RO_AUTH_TOKEN,
   });
   return drizzle(client, {
     logger: process.env.NODE_ENV === "development" ? true : undefined,
