+++ packages/database/drizzle.config.ts
@@ -2,4 +2,5 @@ import type { Config } from "drizzle-kit";
 export default {
   schema: "./schema.ts",
   out: "./drizzle",
+  driver: "turso",
 } satisfies Config;
