+++ packages/database/package.json
@@ -3,7 +3,8 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "sqlite": "drizzle-kit generate:sqlite"
+    "sqlite": "drizzle-kit generate:sqlite",
+    "studio": "drizzle-kit studio --port 3001 --verbose"
   },
   "keywords": [],
   "author": "",
